* libkqueue compiling on windows, but currently untested, broken for sure
git-svn-id: http://opensource.mlba-team.de/svn/xdispatch/trunk@73 9e4c620c-c391-4b63-a7fb-4924991ea895

+++ libkqueue/src/common/debug.h
@@ -27,7 +27,7 @@ extern char *DEBUG_IDENT;
 #elif defined(__sun)
 # define THREAD_ID (pthread_self())
 #elif defined(_WIN32)
-# define THREAD_ID (GetCurrentThreadId())
+# define THREAD_ID (int)(GetCurrentThreadId())
 #else 
 # error Unsupported platform
 #endif
